Refactor, merge files.
