--- .github/workflows/check_construction_native.yml
@@ -1,33 +1,0 @@
-name: Check construction (native currency)
-
-on:
-  pull_request:
-  workflow_dispatch:
-
-jobs:
-  test:
-    runs-on: ubuntu-latest
-
-    steps:
-      - uses: actions/setup-python@v5
-        with:
-          python-version: 3.11
-
-      - uses: actions/checkout@v4
-
-      - name: Install dependencies
-        run: |
-          pip3 install requests bottle
-          curl -sSfL https://raw.githubusercontent.com/coinbase/mesh-cli/master/scripts/install.sh | sh -s -- -b "$HOME/.local/bin"
-          echo "$HOME/.local/bin" >> $GITHUB_PATH
-
-      - name: Build
-        run: |
-          cd $GITHUB_WORKSPACE/cmd/rosetta && go build .
-          cd $GITHUB_WORKSPACE/systemtests && go build ./proxyToObserverAdapter.go
-
-      - name: Run scenario
-        run: |
-          PYTHONPATH=. python3 ./systemtests/check_construction_native.py --network testnet
-
-
