+++ systemtests/check.py
@@ -1,3 +1,4 @@
+import shutil
 import signal
 import subprocess
 import sys
@@ -5,22 +6,24 @@ import time
 from argparse import ArgumentParser
 from typing import Any
 
+import requests
+
 from systemtests import constants
-from systemtests.check_shared import (get_current_epoch,
-                                      run_proxy_to_observer_adapter)
 from systemtests.config import CONFIGURATIONS, Configuration
 
 
 def main() -> int:
     parser = ArgumentParser()
+    parser.add_argument("--mode", choices=["data", "construction"], required=True)
     parser.add_argument("--network", choices=CONFIGURATIONS.keys(), required=True)
     args = parser.parse_args()
 
+    mode = args.mode
     configuration = CONFIGURATIONS[args.network]
 
     process_rosetta = run_rosetta(configuration)
     process_adapter = run_proxy_to_observer_adapter(configuration)
-    process_checker = run_rosetta_checker(configuration)
+    process_checker = run_rosetta_checker(mode, configuration)
 
     # Handle termination signals
     def signal_handler(sig: Any, frame: Any):
@@ -70,7 +73,27 @@ def run_rosetta(configuration: Configuration):
     return subprocess.Popen(command)
 
 
-def run_rosetta_checker(configuration: Configuration):
+def run_proxy_to_observer_adapter(configuration: Configuration):
+    command = [
+        str(constants.PATH_PROXY_TO_OBSERVER_ADAPTER),
+        f"--proxy={configuration.proxy_url}",
+        f"--shard={configuration.network_shard}",
+        f"--sleep={constants.ADAPTER_DELAY_IN_MILLISECONDS}"
+    ]
+
+    return subprocess.Popen(command)
+
+
+def run_rosetta_checker(mode: str, configuration: Configuration):
+    if mode == "data":
+        return run_rosetta_checker_with_check_data(configuration)
+    elif mode == "construction":
+        return run_rosetta_checker_with_check_construction(configuration)
+    else:
+        raise ValueError(f"Unknown mode: {mode}")
+
+
+def run_rosetta_checker_with_check_construction(configuration: Configuration):
     """
     E.g.
 
@@ -87,6 +110,44 @@ def run_rosetta_checker(configuration: Configuration):
     ]
 
     return subprocess.Popen(command)
+
+
+def run_rosetta_checker_with_check_data(configuration: Configuration):
+    """
+    E.g.
+
+    rosetta-cli check:data --configuration-file devnet-data.json \
+        --online-url=http://localhost:7091 --data-dir=devnet-data
+    """
+
+    shutil.rmtree(configuration.check_data_directory, ignore_errors=True)
+
+    current_epoch = get_current_epoch(configuration)
+    first_historical_epoch = current_epoch - configuration.num_historical_epochs + 1
+    start_block = get_start_of_epoch(configuration, first_historical_epoch)
+
+    command = [
+        "rosetta-cli",
+        "check:data",
+        f"--configuration-file={configuration.check_data_configuration_file}",
+        f"--online-url=http://localhost:{constants.PORT_ROSETTA}",
+        f"--data-dir={configuration.check_data_directory}",
+        f"--start-block={start_block}"
+    ]
+
+    return subprocess.Popen(command)
+
+
+def get_current_epoch(configuration: Configuration) -> int:
+    response = requests.get(f"{configuration.proxy_url}/network/status/{configuration.network_shard}")
+    response.raise_for_status()
+    return response.json()["data"]["status"]["erd_epoch_number"]
+
+
+def get_start_of_epoch(configuration: Configuration, epoch: int) -> int:
+    response = requests.get(f"{configuration.proxy_url}/network/epoch-start/{configuration.network_shard}/by-epoch/{epoch}")
+    response.raise_for_status()
+    return response.json()["data"]["epochStart"]["nonce"]
 
 
 if __name__ == "__main__":
